+++ src/main/java/eu/trentorise/opendata/commons/TodUtils.java
@@ -24,6 +24,8 @@ import com.google.common.collect.Multimap;
 import com.google.common.collect.Table;
 import com.google.common.collect.Table.Cell;
 
+import exceptions.TodParseException;
+
 import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
 import java.net.URL;