+++ src/main/java/eu/trentorise/opendata/commons/validation/Preconditions.java
@@ -20,8 +20,9 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.collect.Iterables;
 
-import eu.trentorise.opendata.commons.TodParseException;
 import eu.trentorise.opendata.commons.TodUtils;
+import exceptions.TodParseException;
+
 import javax.annotation.Nullable;
 
 /**